requirements.txt
python-dotenv>=1.0.1 # load_dotenv in CLI and QA chain

typer>=0.12.0 # CLI commands in cqia/cli/main.py

rich>=13.7.1 # console tables/metrics

requests>=2.32.0 # GitHub API client and web calls

fastapi>=0.111.0 # web.service FastAPI app

uvicorn[standard]>=0.30.0 # serve-api command

streamlit>=1.36.0 # web/ui_app.py Streamlit UI

pydantic>=2.6.0 # models in web.service and reporting.schema

pathspec>=0.12.1 # optional .gitignore matching in walker.py

networkx>=3.2.1 # dependency graph build/exports

langchain>=0.2.12 # core message/types used via langchain_core imports

langchain-core>=0.2.38 # direct imports in code

langgraph>=0.2.35 # agent graph flow

langchain-groq>=0.2.0 # ChatGroq integration in qa.chain

groq>=0.11.0 # underlying Groq Python SDK used by langchain-groq

# chroma-hnswlib

chromadb>=0.5.3 # vector DB used by CodeEmbeddingManager (Chroma)

langchain-chroma>=0.2.2 # LangChain’s Chroma integration module

sentence-transformers>=2.7.0 # embedding model all-MiniLM-L6-v2 in CLI index

numpy>=1.26.2,<2.0.0 # compatible with langchain-chroma constraint

scikit-learn>=1.3.0 # used by sentence-transformers dependency tree

torch>=2.2.0 # required backend for sentence-transformers

tree-sitter>=0.21.0

tree-sitter-python>=0.25.0

tree-sitter-javascript>=0.21.0

tree-sitter-typescript>=0.21.0

PyYAML>=6.0.1 # presets/rules.yaml load/save

markdown-it-py>=3.0.0 # safe Markdown rendering fallback for report processing if needed

langchain_huggingface